поменял intval на floatval у фильтра цены https://jira.napoleonit.team/browse/LOKI-241

+++ application/modules/Methods/Model/Methods.php
@@ -520,7 +520,7 @@ class Methods extends Model
                 }
                 // Цена
                 if ( isset($filter['search']['price']) && is_numeric($filter['search']['price']) ) {
-                    $price = intval($filter['search']['price']);
+                    $price = floatval($filter['search']['price']);
                     $where .= "m.price = {$price} AND ";
                 }
